[wip][fix] fix the return type of parse function (u8 -> usize)

+++ src/cpu/decode.rs
@@ -6,9 +6,9 @@ use super::instruction::{OpecodeKind, Instruction};
 pub trait Decode {
     fn decode(&self) -> Instruction;
     fn parse_opecode(&self) -> Result<OpecodeKind, &'static str>;
-    fn parse_rd(&self,  opkind: &OpecodeKind) -> Option<u8>;
-    fn parse_rs1(&self, opkind: &OpecodeKind) -> Option<u8>;
-    fn parse_rs2(&self, opkind: &OpecodeKind) -> Option<u8>;
+    fn parse_rd(&self,  opkind: &OpecodeKind) -> Option<usize>;
+    fn parse_rs1(&self, opkind: &OpecodeKind) -> Option<usize>;
+    fn parse_rs2(&self, opkind: &OpecodeKind) -> Option<usize>;
     fn parse_imm(&self, opkind: &OpecodeKind) -> Option<i32>;
 }
